Switch to BlockedNumberProvider in N SDK
1. ID in BlockedNumberProvider is of type "long" rather than "int".
2. BlockedNumberProvider doesn't include countryIso, so we need to
   calculate it when needed.
3. ag/850887 will remove STRIPPED_PHONE_NUMBER column and notify
   changes to content.
4. Remove unused code.

Bug: 26453530

+++ src/com/android/contacts/callblocking/BlockNumberDialogFragment.java
@@ -82,7 +82,7 @@ public class BlockNumberDialogFragment extends DialogFragment {
     private Callback mCallback;
 
     public static void show(
-            Integer blockId,
+            Long blockId,
             String number,
             String countryIso,
             String displayNumber,
@@ -97,7 +97,7 @@ public class BlockNumberDialogFragment extends DialogFragment {
     }
 
     private static BlockNumberDialogFragment newInstance(
-            Integer blockId,
+            Long blockId,
             String number,
             String countryIso,
             String displayNumber,
